requirements/base.txt
Django>=1.10.6
django-environ>=0.4.1
django-braces>=1.11.0
django-crispy-forms>=1.6.1
django-admin-bootstrapped>=2.5.7
django-authtools>=1.5.0
djangorestframework>=3.5.4
easy-thumbnails>=2.3
django-tables2>=1.4.1
django-extensions>=1.7.7
django-timezone-field>=2.0
olefile>=0.44
Pillow>=4.0.0
pytz>=2016.10
six>=1.10.0
sqlparse>=0.2.3
requests>=2.13.0
PyPrintful>=1.0.0
django-filter>=1.0.1
django-formtools>=2.0
-e git+https://github.com/559Labs/django-crispy-unforms.git@b51040724023a4eb89462840d3dfeb87df63d67f#egg=django_crispy_unforms
